refactor(cli): move fs module under project folder
