+++ crates/tsz-cli/src/lib.rs
@@ -8,15 +8,16 @@ pub mod args;
 pub mod build;
 pub use tsz::config;
 pub mod driver;
-pub mod fs;
 pub mod help;
 pub mod incremental;
 pub mod locale;
+pub mod project;
 pub mod project_refs;
 pub mod reporter;
 pub mod trace;
 pub mod tracing_config;
 pub mod watch;
+pub use project::fs;
 
 #[cfg(test)]
 #[path = "../tests/args_tests.rs"]
